assertion-ingestion-demo/requirements.txt
datahub>=0.12.0
boto3>=1.26.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
